Use Hide Action struct to resemble built-in Dismiss Action
